requirements.txt
pyttsx3==2.90
speechrecognition==3.8.1
wikipedia==1.4.0
pywhatkit==5.1
mtranslate==1.8
openai==0.28.0
requests==2.31.0
pillow==10.0.0
pyautogui==0.9.53
pyaudio==0.2.11